requirements.txt
streamlit
google.generativeai
PIL